requirements.txt
fastapi==0.95.1
faiss-cpu
langchain==0.0.153
pandas==1.4.4
python-ulid 
python-multipart
pydantic
transformers==4.28.1
uvicorn==0.22.0
sentencepiece==0.1.99
InstructorEmbedding==1.0.0
sentence_transformers==2.2.2
openai==0.27.6
python-dotenv==1.0.0
pypdf==3.8.1
python-docx==0.8.11
pyTelegramBotAPI==4.8.0
gdown